A (hopefully) final rework of Geometry::merge() functionality
This fixes any bugs caused by generating/changing data after the merge
that is lost because the merged record sometimes used new GeomData
entries which are not kept. Now merges always use existing records
(fixed or varying) and interleave their data with the merged record.
Record with the vertices is the parent (merge recipient/destination) and
retains the changes.
- First generate a list of all fixed records
- Loop through varying data. If vertices found in matching fixed data,
that is the base record, otherwise use the varying data as the base
(note: this means a fixed vertex record with some varying data is still
marked as timestep -1 and appears as fixed, if this causes problems down
the track will have to add a separate flag)
- When merging values, if data exists under label then replace it
instead of the usual append
- When merging render data, if data found in source record,
overwrite/replace in the destination record
- Fixed records are cleared from fixed list as each is merged
- Finally, any remaining fixed records are copied as is