Render order and custom rendering mega-update
- Full control over renderer creation and draw order
- Renderers created as data is loaded, and called in creation order so
draw order can now be controlled by the order data is loaded into
objects.
- Objects still use a single default renderer for their type as before,
but the order can be controlled without setting "renderlist"
- Setting "renderlist" is still supported to specify the creation order
of default renderers.
- Renderer to use can be controlled by setting renderer="name", where
name is one of the base types or aliases, this replaces the combined
function of the previous "renderer" and "geometry" properties.
- Passing a custom renderer="name" creates a new renderer that is called
separately, but can be shared with other objects by using the same name.
- Custom renderers can have a specific base type by passing
renderer="name:base", otherwise the default for the loading function is
used
- Aliases for data types match to the master list of renderer types,
which then map to the base type.
- Custom shaders can be loaded from source strings instead of filenames
- Timestep now stores optional "time" value in as a property. Added a
time() function which returns the property if available, or just the
step index otherwise.
- List of created renderers available with "list render" or
lv.list("render")
- Shapes renderers improved and fixed, support lv.spheres(), lv.cubes()
etc