Fix #72 and add property allowing old blending style to be used
This bug was caused by an error in the shader where premultiplied alpha
is required, but the result was not always undesireable, in two cases:
- where transparent colour was black, blended smoothly into the
neighbouring colour without showing any dark shading/edges, in the
corrected version the blend is increasingly dark, would need to use the
same rgb as neighbour to achieve similar result.
- where transparent colours are bright, created a brightnening, bloom
style effect in the blend region

In order to preserve this, added the "bloom" property, range [0,1]
default=0
Setting to approximately 0.5 replicates the old blending style and
allows existing applications using colour maps designed for this
blending to be supported.