Fix video encoding in wheel builds
Version of ffmpeg available for redhat/almalinux based manylinux_2_28
image links to libpython6, this breaks the wheels so we have to build
our own ffmpeg, building a minimal version and uploading to the release
in the publish step, then the wheel deploy jobs can use this.
Also switching to packaged version of libOSMesa instead of building our
own.